Fix CV flag
The flag of Cape Verde had the stripes and stars positioned too low.
The top blue part is supposed to occupy half of the height of the flag.
I've edited the geometry so that the white and red stripes start at 50%
of height of the circle. I moved the stars upwards accordingly.

See https://en.wikipedia.org/wiki/Flag_of_Cape_Verde for details.